Fix an issue in Model_User::checkPermission()
There was a typo in the function name and the way it was called as well. Usually permissions are checked through the \Nos\User\Permission class, however this class works only if the user is logged in the admin.
Any application using Model_User as front-end user model would not be able to use \Nos\User\Permission and would have to rely on this buggy method call.