Bugfix for virtualpath
The virtual path was changed for all items without take in count the context. So you could change a url for a subpage "C" of a context by changing the virtualname of a parent page "A" of an other context (if the parent page "B" of the first context has the same virtual name than "A").

A simple example :
A is in context main.
B and C are in context second.
A and B have the same virtual name "ok.html".
C is the child of B.
You change A virtual name to "test".
C virtual name become "test/C.html" instead of staying "ok/C.html".