Fix for plural i18n
The previous fix on plural i18n was not complete, this fix makes the plural work as expected.

The translate_from_file() method return the singular form if no translation were found and if the default value is null, but for the plural() method to work we need to pass a falsy value or an array. To bypass this behaviour we need to set false as default value for translate_from_file().